fix(windows): refuse Arm64 at install; checked backup; honest rollback (#659)
`ArchitecturesAllowed` x64compatible → x64os. `x64compatible` also matches
Arm64 Windows 11, which runs x64 binaries under emulation: the installer
succeeded there and the input method then did nothing in every Arm64-native
application (Notepad, Edge, Office) while working in emulated ones. A text
service is loaded in the host process's own architecture and no Arm64
service is built, so this was an install that looked fine and left most of
the machine without an input method. USER 2026-09-01: no Arm64 release for
now, revisit on a user report.

32-bit hosts are a deliberate gap for the same round (USER 2026-09-01):
everywhere Taiwanese is typically typed is 64-bit, the known exception being
Office 2016 and earlier. `windows-release.md` now carries the measured
record — the one compile fix i686 needs, why the second service must sit
BESIDE the first rather than in a subdirectory, and how the registrations
interact — so the round that ships it does not re-derive any of it.

Installer correctness, found by Codex and independent of architecture:

* The upgrade backup used an unchecked `CopyFile` and unregistered the old
  service regardless. A backup that never happened left the rollback with
  nothing to restore — neither the old service nor the new one. The copy is
  checked now and happens before anything is unregistered; a failure stops
  the install while the machine is still whole, and a restore that fails is
  logged.
* The comment on `RegisterEverything` claimed the exception makes Setup
  "undo its file installation". It does not: Setup finalizes the uninstall
  log before `ssPostInstall`, and its own documentation says errors after
  that point do not roll back what was installed. What RollBack restores is
  the previous service and its registration; the comment says so now.

`ContextToken` is a `usize`: its other life is a `WPARAM` in the window's
queued hide request, so matching the transport's width makes the round trip
cast-free. Its counter in the text service gets the same
stop-rather-than-wrap guard the coordinator's allocator already had — `0` is
the sentinel those messages read as "no token".

Release gates: the C-runtime import check covers what its message claims
(`msvcrt`, `msvcr*`, `ucrtbase`, `api-ms-win-crt-*`, not just `vcruntime` /
`msvcp`), and the four entry points `regsvr32` looks for are read back out
of the export table.

`the_stamp_has_the_macos_bundle_version_shape` asserted the patch was
literally `6`; the 3.6.7 bump turned a property of the encoding into a test
of the version number. It reads the patch from the crate version now.

+++ windows/crates/taigi-windows-core/src/composing/coordinator.rs
@@ -37,13 +37,18 @@ use super::manager::ComposingManager;
 /// COM pointer address: a torn-down context leaves its ownership behind and
 /// the allocator can hand the same address to the next one, which would then
 /// inherit the dead one's half-typed composition).
+///
+/// `usize` because the token's other life is as a `WPARAM` in the window's
+/// queued hide request: matching the transport's own width is what makes the
+/// round trip cast-free, and what would keep it honest on a target where
+/// `WPARAM` is narrower than a `u64`.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct ContextToken(pub u64);
+pub struct ContextToken(pub usize);
 
 pub struct ComposingSessionCoordinator {
     manager: ComposingManager,
     current_owner: Option<ContextToken>,
-    next_token: u64,
+    next_token: usize,
 }
 
 impl ComposingSessionCoordinator {
@@ -57,13 +62,13 @@ impl ComposingSessionCoordinator {
 
     /// A fresh, never-used token for a context the shell just met. `0` is
     /// never handed out (it reads as "no token" in queued messages), and the
-    /// counter cannot silently wrap onto a live value: at `u64::MAX` — which
-    /// no process reaches — it stops advancing and the panic is caught at the
-    /// COM boundary like any other.
+    /// counter cannot silently wrap onto a live value: at `usize::MAX` —
+    /// which no process reaches — it stops advancing and the panic is caught
+    /// at the COM boundary like any other.
     pub fn allocate_token(&mut self) -> ContextToken {
         let value = self.next_token;
         assert!(
-            value != 0 && value != u64::MAX,
+            value != 0 && value != usize::MAX,
             "context token space exhausted"
         );
         self.next_token += 1;
